feat: add syscall support to rv64
