+++ examples/rv-foo/Cargo.toml
@@ -7,4 +7,4 @@ edition = "2021"
 
 [dependencies]
 defmt = "0.3"
-#userspace = { path = "../../userspace" }
+userspace = { path = "../../userspace", features = ["rv64"] }
