+++ userspace/Cargo.toml
@@ -6,6 +6,8 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 #
 [features]
+rv64 = []
+cortex_m = []
 
 [dependencies]
 abi = { path = "../abi" }
